pyproject.toml
[tool.poetry]
name = "the-certain-news"
version = "0.1.0"
description = "The Certain News"
authors = ["Yousef Abu Shanab <josephyousef249@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "app" }]
homepage = "https://github.com/youzarsiph/the-certain-news"
repository = "https://github.com/youzarsiph/the-certain-news"
documentation = "https://github.com/youzarsiph/the-certain-news"

[tool.poetry.dependencies]
python = ">=3.10"
channels = ">=4.2.2"
daphne = ">=4.2.0"
django = ">=5.2.1"
django-cors-headers = ">=4.7.0"
django-countries = ">=7.6.1"
django-filter = ">=25.1"
django-phonenumber-field = ">=9.0.8"
djangorestframework = ">=3.16.0"
djoser = ">=2.3.1"
drf-restwind = ">=0.2.0"
markdown = ">=3.8"
phonenumbers = ">=9.0.10"
pygments = ">=2.19.1"
pillow = ">=11.2.1"
python-dotenv = ">=1.1.0"
wagtail = ">=7.0"
wagtail-localize = ">=1.12.1"


[tool.poetry.group.dev.dependencies]
black = ">=25.1.0"
ruff = ">=0.11.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
